fixed payable function invocation

+++ smart_contract/contracts/SimpleAuction.sol
@@ -1,26 +1,27 @@
 // SPDX-License-Identifier: GPL-3.0
 pragma solidity ^0.8.4;
+
 contract SimpleAuction {
     // Parameters of the auction. Times are either
     // absolute unix timestamps (seconds since 1970-01-01)
     // or time periods in seconds.
     address payable public beneficiary;
-    uint public auctionEndTime;
+    uint256 public auctionEndTime;
 
     // Current state of the auction.
     address public highestBidder;
-    uint public highestBid;
+    uint256 public highestBid;
 
     // Allowed withdrawals of previous bids
-    mapping(address => uint) public pendingReturns;
+    mapping(address => uint256) public pendingReturns;
 
     // Set to true at the end, disallows any change.
     // By default initialized to `false`.
     bool ended;
 
     // Events that will be emitted on changes.
-    event HighestBidIncreased(address bidder, uint amount);
-    event AuctionEnded(address winner, uint amount);
+    event HighestBidIncreased(address bidder, uint256 amount);
+    event AuctionEnded(address winner, uint256 amount);
 
     // Errors that describe failures.
 
@@ -32,7 +33,7 @@ contract SimpleAuction {
     /// The auction has already ended.
     error AuctionAlreadyEnded();
     /// There is already a higher or equal bid.
-    error BidNotHighEnough(uint highestBid);
+    error BidNotHighEnough(uint256 highestBid);
     /// The auction has not ended yet.
     error AuctionNotYetEnded();
     /// The function auctionEnd has already been called.
@@ -41,10 +42,7 @@ contract SimpleAuction {
     /// Create a simple auction with `biddingTime`
     /// seconds bidding time on behalf of the
     /// beneficiary address `beneficiaryAddress`.
-    constructor(
-        uint biddingTime,
-        address payable beneficiaryAddress
-    ) {
+    constructor(uint256 biddingTime, address payable beneficiaryAddress) {
         beneficiary = beneficiaryAddress;
         auctionEndTime = block.timestamp + biddingTime;
     }
@@ -62,16 +60,14 @@ contract SimpleAuction {
 
         // Revert the call if the bidding
         // period is over.
-        if (block.timestamp > auctionEndTime)
-            revert AuctionAlreadyEnded();
+        if (block.timestamp > auctionEndTime) revert AuctionAlreadyEnded();
 
         // If the bid is not higher, send the
         // money back (the revert statement
         // will revert all changes in this
         // function execution including
         // it having received the money).
-        if (msg.value <= highestBid)
-            revert BidNotHighEnough(highestBid);
+        if (msg.value <= highestBid) revert BidNotHighEnough(highestBid);
 
         if (highestBid != 0) {
             // Sending back the money by simply using
@@ -88,7 +84,7 @@ contract SimpleAuction {
 
     /// Withdraw a bid that was overbid.
     function withdraw() external returns (bool) {
-        uint amount = pendingReturns[msg.sender];
+        uint256 amount = pendingReturns[msg.sender];
         if (amount > 0) {
             // It is important to set this to zero because the recipient
             // can call this function again as part of the receiving call
@@ -124,10 +120,8 @@ contract SimpleAuction {
         // external contracts.
 
         // 1. Conditions
-        if (block.timestamp < auctionEndTime)
-            revert AuctionNotYetEnded();
-        if (ended)
-            revert AuctionEndAlreadyCalled();
+        if (block.timestamp < auctionEndTime) revert AuctionNotYetEnded();
+        if (ended) revert AuctionEndAlreadyCalled();
 
         // 2. Effects
         ended = true;
